feat: Generate constant random byte arrays
The `const_random` macro will now accept and generate fixed sized `u8`
arrays of any size.

Implementation detail: This patch replaces the 256 bit fixed sized sha3
hash (sha3-256) with the sha3 eXtendable-Output Function (SHAKE256).
This is to allow
arbitrarily long byte arrays to be generated. This is not strictly
necessary because `tiny-keccak` does not limit the output size of sha3
functions but I feel that misusing crypto primitive is a code smell.